backend/requirements.txt
asgiref==3.9.0
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
psycopg2-binary==2.9.10
redis==5.0.1
django-redis==5.4.0
django-ratelimit==4.1.0
cryptography==41.0.7
django-axes==6.1.1
